// types.ts
export type LibraryCategory = 'Past Papers' | 'Textbooks';

export interface LibraryDocument {
  id: string;
  name: string;
  category: LibraryCategory;
  uploadDate: string;
  dataUrl: string; // Base64 string for storage or Object URL
  size: string;
  type: string; // MIME type
  isDefault?: boolean;
}